fixed TRACK mode data storage!
turns out, I forgot that scanf() and sscanf() have to use the pointers of data in order to store them. so now we've got the right length of trajectory and can do something with it

a problem I'm currently having: the first and last datapoints of the data sent back don't seem to match the ref array. I'll work on that next