major modularity adjustments
Nick suggested that I make sure to define variables in the functions that mainly use them. this helps both the modularity of the code, and makes it easier to debug, since variables are defined and referred to in fewer places

the compiler got mad when variables were defined in the .h files, so global variables are defined in .c files now. if main() needs variables, we're using extern to find them, rather than have them in the header