same errors, took some screenshots
still trying to get a way to solve this serial issue. the amount of data being sent by the PIC - this is confirmable in PUTTY - stops being real data at a certain point, and just pauses

this is confusing b/c just manually inputting a trajectory m and having the PIC spit it out works fine. it's just: the PIC craps out after 1 round of the trajectory testing